Implement semantic analysis of nested select statements in FROM clause.
Remove the `name` attribute from the FROM clause `from_type` struct.

Rename Sema Context `named_expr` into `results`.

The duty of pushing and popping a sema context is now on the call site
(e.g. the FROM clause or the WHERE clause), and not the callee (i.e.
when the visitor reaches a SelectStmt).

Allocate `SemaContext` on the heap instead of on the stack.  This allows
us to give persistent handles to the contexts and avoids nasty bugs
caused by `std::vector`s "realloc".