Progress on the WASM backend.  (Squashed commit)
Move backend factory method definitions into respective source files.
Get rid of 'Backend.cpp'.

Introduce memory file-backed allocators.

Introduce an interface for custom allocators with memory file backed
memory.  This is required to enable the rewiring technique.
Implement a linear allocator with that interface.

Add a linear allocator to the `Catalog`.

This allocator will serve as our allocator for memory file backed
allocations.  It must be used to enable rewiring.  This is necessary to
implement WASM modules, since they are limited to 2 GiB memory.

Modify `RowStore` and `ColumnStore` to use the `Catalog`'s memory
allocator.

Add a macro to explicitly discard the return value.

Some functions in V8 are marked to warn if the return value is not used.
Define a macro to mark the return value as explicitly discarded to
silence this warning.

Introduce a `WasmContext` to the `WasmPlatform`.

The `WasmContext` tracks information belonging to a running WASM module
instance, e.g. the underlying virtual address space, and associates it
with the module instance's unique ID.  This unique ID must be the first
parameter to the invoked module's function.

Don't import the memory into the wasm module.

It is not necessary to create a `WebAssembly.Memory` instance and import
it into our `WebAssembly.Instance` instance.  A call to
`v8::SetWasmInstanceRawMemory` is enough to set the underlying memory.

Switch from Google V8 to our fork of V8.

Include a small micro benchmark for WASM calls.