Add helper method `Sema::make_designator()`.
This helper method creates a new `Designator` that is located at the
same `Position` as `from`.  It is an identifier (i.e. without table
name) with the attribute name set to the string representation of `to`.
Its `type_` is set to that of `to` and its `target_` is `to`.

This method can be used to replace an `Expr` subtree by a single
`Designator` in case the subtree is already computed in an earlier
stage.