Implement using expressions as grouping keys.
So far, expressions were allowed as grouping keys, but the expressions
were re-evaluated in the later stages after grouping.  This is often
redundant and can also lead to the construction of invalid plans.

Instead, use the new `Sema::make_designator()` method to replace the
uses of grouping keys in later stages by `Designator`s pointing to the
grouping key in the `GroupByClause`.  This rewrite of the AST resolves
both redundant re-evaluation and the construction of invalid plans.