Implement `DotTool` to render and show dot output.
The `DotTool` class dynamically loads the `graphviz` shared library to
render our dot output directly to PDFs.  If the `graphviz` library
cannot be dynamically loaded (via `dlopen`), it falls back to writing
the dot to a file.  The `DotTool` provides the very convenient function
`show(char *name, bool interactive)` that chooses the best way to
present the graph to the user.