Fix `nullptr` dereference if `TERM` not in env.
In `term::has_color()`, we use `std::getenv("TERM")` to get the name of
the terminal.  The call may return `nullptr` if `TERM` is not set in the
environment.  This causes a `nullptr` dereference on the CI runner.
With this commit we handle the case where `getenv` returns `nullptr`
gracefully and simply assume no color support.