Provide new `Value` and `Tuple` classes.
- The new class `Value` represents a SQL value of any type, e.g. `42` of
type `INT(4)` (aka i32) or `3.14` of type `float`.  However, it cannot
represent `NULL`.  `Value` is implemented as a simple C `union` with
some getters/setters for easy and sane access.  In `Debug` build,
`Value` stores its current type and accesses validate correctness of the
assumed type.  In the `Release` build, no such type information exists.

- The class `Tuple` represents a sequence of fixed, finite length of
`Value`s.  Additionally to the `Value`s, a `Tuple` has a `NULL` bitmap
to mark slots in the `Tuple` as `NULL`.  The class provides
getters/setters for easy access and adds some `NULL` sanity checks.  In
the `Debug` build, a `Tuple` stores its size and accesses are checked
for out-of-bounds.  In the `Release` build, no size information is
stored and no such checks are performed.

Adapt the implementations of `Interpreter` and `StackMachine`
accordingly.