Implement `DPccp`.
Address issue #32. Implement the `PlanEnumerator` algorithm `DPccp` as proposed in
'Analysis of two existing and one new dynamic programming algorithm for
the generation of optimal bushy join trees without cross products' by
Moerkotte and Neumann.  In contrast to the algorithm described in the
paper, we use a queue to avoid excessive recursion.