Fix #29 Adjust float precision in PostgreSQL setup
In PostgreSQL, `FLOAT` defaults to double-precision resulting in
different results for arithmetic operations.  We update the setup
script for PostgreSQL to use `REAL` instead, which is internally
represented by a 32-bit variable-precision foating point.  We add an
integration test case to verify that the precision of mutable and
PostgreSQL are now matching.