Fix #35
`minimize_schema` computes the `OperatorSchema` for `JoinOperator`s
using `get_attr_ids()`.  In case the given join predicate contains a
specific `Designator` multiple times, the corresponding
`AttributeIdentifier` is also added to the `OperatorSchema` multiple
times.  Subsequently, this leads to an ambiguous lookup for an
`AttributeIdentifier`.  Therefore, when computing the `OperatorSchema`,
we only add a given `AttributeIdentifier` to the schema if it is not
already present.