Change signature of `StackMachine::operator()`.
This commit solves most parts of issue #36.
The evaluation of a `StackMachine` instance with
`StackMachine::operator()` now takes arbitrarily many `Tuple`s as
parameters in the form of an array of pointers `Tuple **tuples`.  The
opcodes are extended as described in issue #36 and the call sites are
adapted to the new interface.

With the new interface, the comparator in `SortingOperator` is
implemented such that it takes the two tuples to compare as input
`Tuple`s `1` and `2` and produces its boolean comparison result in
`Tuple` `0`.