Fix the patch command for replxx.
Our previous patch with `sed` does not apply to a `RelWithDebInfo`
build, because replxx does not event set the output name in that case.
Hence, the library's internal name `replxx-static` is used to produce
`libreplxx-static.a`.  The new patch will set the output name of the
library independent of the build type.