Fix #51: Parse decimals in `DSVReader`.
We first read the part before the dot, then "scale" many places after
the dot, and combine both numbers according to scale. Note that we do
not perform a sanity check on whether the data read from the CSV file
fits into the given data type as disussed in Issue #52.