Fix #50: Emit correct decimal load/store opcodes.
Both `RowStore` and `ColumnStore` had a bug in emitting the appropriate
load `Ld_XX_YY` and store `St_XX_YY` instructions.  We now rely on
`Type::size()` to get the required size in *bytes* to store an instance
of a type.  An important aspect of `size()` is that the size of
`DECIMAL`s is already ceiled to a power of two.