Implement #53: `PlanEnumerator::TDbasic`.
Implement top-down plan enumeration algorithm `TDbasic`.  `TDbasic` is
described in 'Algorithms for Efficient Top-Down Join Enumeration.' by
Pit Fender on pages 26-33. The algorithm uses a generic top-down
enumeration algorithm with naive partitioning. It is a very basic
baseline and can be seen as a starting point for top-down plan
enumeration.