Introduce specialized  `PrintOperator`.
The new `PrintOperator` can be considered a specialized version of
`CallbackOperator` for printing.  This specialization allows us to,
e.g., compile a `StackMachine` instance to perform the printing and
avoid dynamic dispatching.  Hopefully this improves the *observable*
performance of the `Interpreter`.

This new `Operator` also makes way for printing results directly from
the compiled WebAssembly code rather than writing all results to memory,
then reconstructing `Tuple` instances on the host side and invoking the
callback function of a `CallbackOperator` instance.