Add `--debug` flag in 'IntegrationTest.py'.
If `--debug` is provided, the script prints the commands for running
failed tests isolated.  Further, the following changes were made:
- clean up reporting,
- replace `stage` by `stage_name` were applicable,
- repended `num_` to expected and actual values that are numeric.
If `--debug` is provided, the script will print a terminal command for
each failed test case that lets the user run the test cases isolated,
e.g. in a debugger.