Implement script for automated benchmarking.
The script 'benchmark/Benchmark.py' runs benchmark suites in
'benchmark/' and compiles a HTML report with interactive charts.  The
measurements are also collected in a CSV file for later processing.

A benchmark *suite* is composed of multiple benchmarks.  A *benchmark*
consist of one or more YAML `*.yml` files.  Within such a YAML file, a
sequence of *cases* gives the queries to benchmark.  All measurements
produced by the same YAML file make one chart in the report.  Further,
measurements of different YAML files that belong to the same benchmark
are combined in a single chart for easy comparison.

To collect times of queries we provide the `-t` flag to the `shell`
binary and extract the times from the produced output using a regular
expression, that is specified within the YAML files.  This technique
allows the creator of a benchmark to select what time to report, e.g.
time of execution, time of compilation, time of optimization, etc.