Use fresh `Timer` in `process_stream`.
We used a global `Timer` instance to measure times.  This had the
negative side effect of leaving measurement artifacts, e.g. a `SELECT`
query would still show the last time it took to read a DSV file.  We now
use a fresh `Timer` instance for every sequence of commands sent to the
shell.