Fix #71: Implement the `readonly` benchmark flag.
Benchmarks that set `readonly` to true are now evaluated more
efficiently.  All cases are combined and executed on one instance of the
database.  This avoids reloading data for every case.

The timeout per query is set to 5 minutes.  This is the time for both
*loading and executing* the query.  In a readonly benchmark, the timeout
for running *all* cases is the default timeout of 5 minutes plus 30
seconds per case.