Greatly improve time to run benchmarks.
Introduce a new `tables` field in the YAML benchmark files, listing all
tables accessed by the queries of that benchmark.  The `schema.sql` file
no longer contains `IMPORT INTO` statements to load data from the CSV
files into the tables.  Instead, the benchmark script will generate
respective `IMPORT INTO` statements to load data for the accessed
tables.  This saves loading data for tables *not* accessed by that
benchmark.

Along the way, the schema of the `operators` benchmark was restructured
and the system comparison scripts were adapted to the new schema.