Implement an open addressing hash map.
Provide an implementation of a hash map or multimap, respectively, that
uses open addressing with quadratic probing.  More precisely, triangular
numbers are used together with capacities being powers of two.  This
guarantees that probing finds a free slot if one exists.

The hash map uses reference counting to handle duplicates more
efficiently.  It provides interfaces to prohibit or permit the insertion
of duplicates.

Additionally, provide unit tests for the new hash map.