Fix locating the start of data to sort in ORDER BY
The head of heap, that points to the start of the array of elements to
sort, was saved to early.  Later operations, such as hash-based
grouping, would alter the head of the heap and render the previously
saved start of data incorrect.  Instead, delay saving the head of heap
right until the array of data to sort is filled.