Add missing parts of `WasmHashMumur3_64A`.
The implementation of `WasmHashMumur3_64A` was missing an addition of a
constant to the hash inside the loop as well as a final bitwise xor of
the hash and the key length.

Hopefully, this fixes the performance anomaly observed in the benchmark
`group_by_i32_multiple` at 2 attributes to group by.