Add flag to run only required integration tests.
We have a flag to *require all* tests to pass, including those that are
marked as not required to pass.  On the contrary, we had no flag to
*exclude* the tests that are not required to pass from execution.

This commit introduces the `--required-only` (short `-r`) flag to run
only those tests that are required to pass.  This helps debugging
required tests, as no errors from not required tests are shown.