Fix bug in Wasm join code.
The probe phase did not properly handle overflow chains that wrap at the
end of the hash table.  Instead of wrapping the search and continuing at
the beginning of the table, the search proceeded without wrapping.  This
causes reads of uninitialized (and hence all zeroed) memory or -- even
worse -- reading unmapped memory, causing a segmentation violation.