Fix a bug in Wasm HT.
There was a bug in `WasmRefCountingHashTable` in routine
`insert_with_duplicates` where the result of the Gaussian sum was not
properly wrapped.  This led to out-of-bounds accesses and potential
segmentation violations.  In the new approach, we avoid wrapping by
subtraction and instead use an approach that applies the HT's mask.