Fix ignoring `Lienarization` update in `DSVReader`
The `DSVReader` instructed the `Interpreter` to compile the
`Linearization` of the `Store` exactly once before parsing.  However, by
successive calls to `Store::append()`, the `Linearization` might change.
Therefore, the `DSVReader` now checks whether the `Linearization` of the
`Store` changed and then recompiles the `Linearization` to obtain a new
`StackMachine` to write data to the store.