Fix store instructions in `StackMachine`.
The store to memory instructions in the `StackMachine` did not correctly
handle NULL.  If the value to store was NULL, the `StackMachine`
attempted to convert this value to the type to store.  This is a bug
that did not cause any harm until a previous bug in `Tuple` was fixed.
Now, the conversion of the NULL value to the type to store *will always
fail*.  To fix this, the `StackMachine` now simply skips the store
altogether in case the value is NULL.