Fix a bug in storing booleans in the `Interpreter`
The `Interpreter` compiled the `Linearization` in case of a store
incorrectly.  This lead to setting a boolean to FALSE being a no-op.
This is erroneous if the store does not initialize memory or if an
update-statement sets a boolean from TRUE to FALSE.

Our solution is to load the byte where the bit is to be stored and
compute two versions of that byte: one version with the respective bit
set to 1 and one version with the respective bit set to 0.  Then, we
select the correct byte to write back to memory based on the truth value
of the boolean to store.