Revert "Fix type error in `Tuple`."
This reverts commit e3773708ba9b9368eaa41c97c384d7e046529435.

Resetting the `Value` in the `Tuple` is incorrect.  This erases not only
the `Value`s type but also the stored value.  In case of, e.g., a
`CHAR(N)`, this would drop the *only* pointer to the allocated memory
where the `CHAR(N)` is supposed to be stored.

In the meantime, we have resolved the original problem by introducing
additional `NULL` checks in the `StackMachine`.  The `StackMachine` will
only attempt to interpret a `Value` as of a certain type if it is not
`NULL`.