Fix connectedness check in `DPsizeOpt`.
The `PlanEnumerator` `DPsizeOpt` checked the connectedness of a
subproblem by comparing its cost to a numerical constant.  This is not
well defined and may lead to incorrect results in case a plan reaches
the maximum representable cost.  Instead, use the adjacency matrix to
compute to connectedness of a subproblem.
(Note: Alternatively, `PlanTable::has_plan()` can be used.)