Add fast connectedness check to `PlanEnumerator`s.
Rather than using the `AdjacencyMatrix` to compute whether a subproblem
is connected, sometimes a lookup in the `PlanTable` suffices.  If a
subproblem must have been planned already, the `PlanTable` can be used.

As an example, when considering S1 ⋈  S2 in an algorithm for bottom-up
plan enumeration, then S1 and S2 must have been planned already.  Hence,
checking whether there is a plan for S1 is enough to check whether S1 is
connected.  If there is no plan in the `PlanTable`, S1 must be not
connected.  (The same applies to S2.)