Fix bug in Wasm codegen that hindered optimization
Binaryen expects that every `BinaryenExpressionRef` be used at most
once.  However, we did not ensure this and in many places reused an
expression to avoid re-generating code.  However, Binaryen's graph
structure then has multiple uses of the same expression.  This violates
the constraint that Binaryen's graph must be a tree (and not a DAG, as
we thought).

Surprisingly, the code generated *before this fix* passed validation via
`BinaryenModuleValidate()`.  That is, because by default Binaryen
doesn't check for multiple uses of an expression.  Even more
surprisingly, the *incorrectly generated* code that we then handed over
to V8 yielded the expected results when executed.

When running Binaryen's optimization via `BinaryenModuleOptimize()` with
an optimization level greater than 0 caused Binaryen to run into
assertions or crashes.  After long investigation, the source of the
problem can be identified as using a `BinaryenExpressionRef` multple
times.  When optimization is run, an expression might be transformed for
one user, which then also affects the other user of the same expression.
The optimizations simply assumed that the graph is a tree.  Because
Binaryen by default runs optimizations on the IR concurrently, this can
lead to races and UB.

To have Binaryen run more checks during validation and to run
optimization serially, one can set the environment variable
`BINARYEN_PASS_DEBUG=1`.  This greatly helped to identify the cause of
the error.

We have opened an issue with Binaryen to get aid in finding the bug in
our code generation: https://github.com/WebAssembly/binaryen/issues/3532

With this patch, we introduce a new type `WasmTemporary` that captures a
`BinaryenExpressionRef` and ensures the expression is used at most once.
The type `WasmTemporary` has similar semantics to `std::unique_ptr`.
The patch requires to modify almost every line of the code gen to make
sure we always use `WasmTemporary` where we used `BinaryenExpressionRef`
before.  We also upgrade to Binaryen `version_99` to get
`BinaryenExpressionCopy()`, which creates a deep copy of an expression.
This is very useful when we want to avoid re-generating code.

With this patch, Binaryen's optimization runs without errors.