Fix bug in `WasmRefCountingHashTable::rehash()`.
The code to rehash a `WasmRefCountingHashTable` assumed that there is at
least one entry in the table.  While this is usually true, a user might
rehash an empty table on purpose before filling it with data or (as was
the case that made us aware of the bug) an initial capacity of 1
requires a rehashing of the empty table before the first element can be
inserted.