Fix bug in hacky estimation of number of groups.
In the Wasm backend in the codegen for `GroupingOperator` we do a hacky
estimation of the number of groups that will be created to initialize
the hash table.  This estimation had a flaw, where it estimated a single
group in case no estimation was possible.  We now bail out of the
estimation procedure if estimation is not possible and leave the
`initial_capacity` unmodified.  This allows other hacky estimations to
kick in.