Fix composite foreign keys in CB database schema
The tables with composite keys were not referenced by a composite
foreign key, but the parts of the composite key were added as references
individually.  However, this is not possible, as a single part of the
composite key does not uniquely identify a record.

Fix the definition for DBDiagram.io to use composite foreign keys.
Re-generate the schema definition for PostgreSQL.