Implement generic cardinality estimation.
- Define interface `CardinalityEstimator` that is used to assign a
`DataModel` to subproblems.
- A `DataModel` describes the result of evaluating a subproblem.  It is
interpreted by its `CardinalityEstimator` to perform predictions, e.g.
of the cardinality.
- Implement `CartesianProductEstimator` to estimate joins by assuming a
Cartesian product (i.e. join selectivity of 1).
- Implement `InjectionCardinalityEstimator` that allows for hard-coding
cardinalities of subproblems in a separate JSON file.  This enables the
injection of pre-computed cardinalities.
- Implement translation of the query optimization problem to PDDL, such
that a QEP can be calculated with FastDownward.
- The computed `DataModel`s replace the former `size` in the
`PlanTable`.  Now, for every (explored) subproblem, a `DataModel` is
stored to model the result of evaluating that subproblem.
- Add CLI flags to select the preferred cardinality estimator and to
provide the JSON file of pre-defined cardinalities.
- Add unit tests for `InjectionCardinalityEstimator`, testing that the
JSON file format is correctly understood and "prediction" is done
correctly.