Implement simple yet generic visitors.
- Provide a generic `Visitor` base class.  It uses the CRTP design to
get the concrete type of the visitor to visit the class hierarchy with.
- Provide a macro `M_DECLARE_VISITOR(NAME, BASE_CLASS, CLASS_LIST)` to
declare a subclass `NAME` of `Visitor` that visits the class hierarchy
originating at `BASE_CLASS` and containing the subclasses `CLASS_LIST`.
(It is only necessary to provide the non-abstract classes of the class
hierarchy in `CLASS_LIST`; abstract classes are not visitable.)
- Provide a macro `M_MAKE_STL_VISITABLE(VISITOR, BASE_CLASS,
CLASS_LIST)` that generates a `visit()` function in the style of
`std::visit()`.  Also gracefully handles return values in the visitor.
- `M_MAKE_STL_VISITABLE()` is performed autmatically when
`M_DECLARE_VISITOR()` is used.
- There is also `M_DECLARE_VISITOR_WITH_BASE()` to provide a specific
base class for a visitor implementation.  This is useful to dissect an
API visitor interface from an internal (mutable) visitor implementation.