Fix infinite recursion of `InternalStoreVisitor`.
- Drop `M_DECLARE_VISITOR_WITH_BASE()` as it is not possible to extend
the functionality of a visitor.  One can only override existing methods.
The problem here is, that the `accept()` method in the visited class
hierarchy invokes the visitor and by that the compiler performs overload
resolution based on the *base* class of the visitor.  In particular,
only the methods present in the base class take part in overload
resolution.
- Get rid of `InternalStoreVisitor`.  Make `StoreVisitor` an internal
(non-API) class.
- Implement `Store2Wasm` and `WasmStoreCG` as subclasses of
`StoreVisitor`.