Fix two bugs in `AddressSpace`.
The move c'tor of `AddressSpace` did not properly initialize the fields
before swapping.  This may lead to UB when `other` is destructed when
the c'tor returns.

The d'tor of `AddressSpace` did not check whether `addr_` actually
points to mapped memory.  Calling `munmap` with `nullptr` may be illegal
on some platforms.