Fix bugs in coverage report generation.
- With a recent modification of the top-level `CMakeLists.txt`, the
`--coverage` flag set by the `coverage.sh` script was not properly
passed to the compiler.
- The generated `gcov.sh` script -- used to let `lcov` use `llvm-cov
gcov` -- could not be invoked by its relative path.  We now use the
absolute path instead.
- An experimental change is to *not* delete the `.gc*` files --
generated as byproduct of compilation with `--coverage` -- anymore and
to *not* clean the build directory anymore.  This hopefully speeds up
coverage report generation.  A potential negative side effect may be
that updates don't make it into the report.  We have to see...