Fix missing disposal of V8 memory.
We create an `AddressSpace` of roughly 4 GiB per Wasm module and rewire
data into that space.  However, we did not dispose of that space after
execution of the module.  This somehow lead to copy-on-write (triggered
by a later execution) and caused a memory leak.

We now properly dispose of the `AddressSpace` as soon as it is not
needed anymore, i.e. after retrieval of a query's result set.