Support `AddressSpace`s of zero size.
There was a bug in the `AddressSpace` c'tor, where a `size` of 0 (zero)
would call `mmap()` like for any other value of `size`.  However, `mmap`
forbids its `length` parameter (our `size`) to be 0, and raises `EINVAL`
and returns `MAP_FAILED`.  Therefore, we now test whether `size` is 0,
in which case the `AddressSpace` remains default initialized.