Improve checkpoints heuristic.
The checkpoints heuristic calculates a series of checkpoints of
pre-determined sized where each checkpoint must be reachable from its
predecessors.  However, it may happen that a checkpoint of some size *N*
if not reachable from its immediate predecessor.  So far, the heuristic
completely bailed out and fell back to `hsum`.  Instead, we now *skip*
this checkpoint size and proceed with finding the next checkpoint.