Change enum macros.
The `DECLARE_ENUM` and `ENUM_TO_STR` macros assumed a comma `,` between
the elements of the `LIST`.  This, however, limits the use of `LIST` in
other macros where there *must not* be a comma in between elements.  The
new definitions insert a comma between elements, such that the original
definition of `LIST` can be free of commas.