Fix `difference_type` in `doubly_linked_list`.
The `difference_type` was an unsigned type.  This lead to unexpected
behavior of STL routines, such as `std::advance()` or `std::prev`.
These routines use `difference_type(-n)` to decrease an iterator by `n`
steps.  The cast would wrap, since `difference_type` was an unsigned
type, leading to a vast increase of the iterator, most likely going
out of bounds and leading to UB.
This is fixed by using a proper signed type for `difference_type`, here
`std::ptrdiff_t`.