Implement a list-based memory allocator.
The `list_allocator` maintains a list of unused memory chunks to avoid
aggressive deallocation and to serve future allocations with already
available memory.  It supports splitting large chunks to avoid waste
when serve small allocations as well as coalescing deallocations to
reclaim large chunks and to reduce fragmentation.  The underlying linked
list is maintained sorted by the chunks' addresses.  The allocator
allows for specifying an initial allocation size (for pre-allocation)
and an allocation strategy (currently linear or exponential).

The allocator's runtime behavior depends on the allocation/deallocation
pattern as well as on the allocation strategy.
For *n* allocations, the allocator exhibits *O(n)* worst-case runtime,
as for each allocation or deallocation the entire list must be
traversed.  Using the exponential allocation strategy, the list's length
is capped to *O(log n)* and hence the worst-case runtime is *(O n log
n)*.