Update Binaryen to v102.
We had to update Binaryen because of the following error, that occurs
when compiling with Clang v13 in combination to Binaryen's `-Werror`
setting
```
binaryen/src/wasm-type.h:358:10: error: definition of implicit copy constructor for 'Tuple' is deprecated because it has a user-declared copy assignment operator [-Werror,-Wdeprecated-copy]
```

Additionally, the signature of `BinaryenSetMemory()` changed to take a
`bool*` instead of `int8_t*` parameter.  We adapt the calling code
accordingly.