Implement `SubsetEnumerator`.
The `SubsetEnumerator` enumerates all subsets of a given size of a given
`SmallBitset`.  For example, if the given `SmallBitset` is `0b1101` and
the size is `2`, then the `SubsetEnumerator` will enumerate `0b0101`,
`0b1001`, and `0b1100`.

`SubsetEnumerator` builds atop `GospersHack` and uses it internally.
However, note that `GospersHack` enumerates subsets solely based on the
size of the set and subset.  It does not acknowledge the elements within
the set.  More precisely, `GospersHack` cannot be used to enumerate the
subsets of a `SmallBitset` with a *gap* in between, like `0b1101`.