Implement a tool to generate fake cardinalities.
Our tool `cardinality_gen` can be used to generate fake cardinalities
for a query.  More precisely, the tool generates cardinalities for all
relations and all feasible subproblems (subset of relations that are
connected by joins).  As output, the tool produces JSON that can be
passed to our `InjectionCardinalityEstimator` and hence the generated
cardinalities can easily be injected into mu*t*able.

TODO: We generate cardinalities of relations and selectivities of joins
using a Gamma distribution, that is derived from parameterized Gamma
function.  We hand-picked these parameters.  Perhaps there is a better
choice for these parameters to achieve more desirable distributions of
cardinalities and selectivities.