Improve `InjectionCardinalityEstimator`.
In the `DataModel`, represent joined relations no more as `std::vector`
of `const char*`.  Instead, we now use the `SmallBitset` encoding in the
`DataModel`.  Given the `QueryGraph`, we can compute the identifier
(that is composed of the individual `DataSource`s' names) when needed.
This saves a lot of memory and operations to estimate a binary join.  It
further enables us to support joins of multiple subproblems at once.