Rework the checkpoints heuristic.
This commit completely overhauls the checkpoints heuristic.  The core
idea remains unchanged (and hence we stick with the name):  The
checkpoints heuristic calculates for a given state a sequence of
checkpoints, such that a path through these checkpoints is likely a
cheap path to the goal.  Once the checkpoints have been computed, the
heuristic plans the path from the given state to the goal by solving the
smaller planning stages from given state to first checkpoint, from first
checkpoint to second checkpoint, and so on, until it eventually plans
the path from the last checkpoint to the goal.  The heuristic value for
the given state is then the sum of the costs for the individual planning
stages between the checkpoints.

With this commit, we extend the heuristic to consider many more
checkpoints while at the same time reducing the overall computation
costs.  Instead of pre-calculating fixed sets of checkpoints and
predicting their cardinalities, the heuristic now explores for the given
state the connected subproblems that are reachable from this state by
joining *N* subproblems of the given state, where *N* is a constant
currently set to 3.  Of all these connected subproblems, the heuristic
chooses the one of least cardinality, then substitutes this one for the
subproblems that it is composed of, and repeats until no next checkpoint
of *N* subproblems exists.

This approach remedies the necessity to pre-compute a potentially
unnecessarily vast amount of checkpoints (depending on the query
topology).  Furthermore, by considering more checkpoints, the heuristic
improves in accuracy and works particularly much better on chain and
cycle tolopogies than the old checkpoints heuristic.