Fix `BaseTable` creation in `GraphBuilder`.
The `GraphBuilder` created `BaseTable`s always with an alias.  If no
alias was specified in the query, the underlying `Table`'s name was used
as alias.

The later introduced method `BaseTable::name()` provides a name lookup,
that considers an alias only when present.  Old method `alias()` simply
returns the alias, or `nullptr` if no alias is present.

By fixing the construction of `BaseTable` in `GraphBuilder`,
consequential errors emerged and had to be fixed.  One must carefully
choose between `alias()` or `name()`, depending on the information one
is interested in.