Implement IK/KBZ plan enumeration.
Implement the IK/KBZ algorithm for computing an optimal linear plan. The
algorithm is a combination of two works:

- Toshihide (I)baraki and Tiko (K)ameda. "On the optimal nesting order
  for computing n-relational joins."
- Ravi (K)rishnamurthy, Haran (B)oral, and Carlo (Z)aniolo.
  "Optimization of Nonrecursive Queries."

Note that the `rank()` is defined according to the cost model.  The
current implementation assumes the `SimpleCostFunction` model to be used
in optimization.  Using another cost model requires adapting the
`rank()` in the IK/KBZ algorithm.  If the `rank()` is not adapted to the
cost model, the linear plan found is no longer guaranteed to be an *optimal
linear plan* for the query.