Fix includes in `QueryGraphTest.cpp`.
There was an include of a **source** file `.cpp`, which is forbidden.¹
The reason to include that source file was to get access to functions
not declared in the respective header.  The solution here is to
manually forward-declare these functions in `QueryGraphTest.cpp` (refer
to this commit's changes).

(¹) There is one exception to the rule: if a source file `.cpp`
**implements** templated functions for methods that have only been
**declared** in a header file `.hpp`, then the source file can be
included to provide the implementations of the templated
functions/methods.  This technique allows for separating template
function/method declaration from definition.  It can further be combined
with explicit template instantiation.