Fixed overflow in `LinearSpace`.
When calculating the final step of a full range `int64_t`, the cast from `double` to `int64_t` would cause an overflow. To prevent this, `step_` is now calculated with a floating-point rounding direction towards zero. This causes some imprecision in `LinearSpace` for large values but prevents this overflow.
Added extra unittests to test `LinearSpace` at full range for `int32_t` and `int64_t`.