Introduce method `for_each_successor()`.
The search states now provide a method `for_each_successor()` to allow a
caller to iterate over each subproblem within the state via a callback.
Simultaneously, a search state needs no longer provide an iterator
interface to the subproblems.  This is useful for states that do not
store the subproblems but compute them on demand.