Fix `compute_datasource_to_subproblem_index()`
There was a bug in
`SearchStateEdgesBottomUp::compute_datasource_to_subproblem_index()`
where not all necessary entries of the reverse index were updated but
only one.  Instead of updating only the entry for the right sources in
the join, we must update the entries for all relations in the subproblem
containing the right source.