`eval-planners.sh` now skips timed out configs.
Previously, the script used an associative array to map pairs of a
topology and a planner configuration to the maximum number of relations
still supported.  This array was then used to skip over too large
problems to speed up evaluation.

The problem is, that this array must be maintained and the values result
from actual experience.  With new topologies and/or planners, the array
must be extended.  This does not scale well and is not very portable
(because of hardware differences).

Therefore, the script now stores occurrences of timeouts for
topology-planner pairs.  If a certain threshold of max. timeouts is
reached, the planner config is skipped for all further experiments with
that topology.  Currently, that threshold is set to 1.