Implement search state partitioning.
States in the search space can now provide a partitioning.  A static
method `num_partitions()` gives the number of total partitions and
method `partition_id()` returns the partition ID of a state.

Partitioning is used to keep states of different partitions in different
hash tables.  This speeds up duplicate checking and rehashing.