Support both shared and static library build.
With this major patch, mu*t*able can now be built with either shared or
static libraries.  In a shared library build, all dependences are built
as shared libraries, too, and dynamically linked to all executables.
This saves a lot of storage space and enables fast recompilation.  In a
static library build, all third parties as well as `libmutable` are
linked into the executables.  This allows for removing unused symbols
and easy shipping of single executables.

To enable both static and dynamic library builds, several steps were
necessary:

- Refactor the `CMakeLists.txt` of the `src/` directory and the
  `backend/` subdirectory.
- Add specific linkage flags in a static build to not discard unused
  symbols when linking to `libmutable.a`.
- Refactor the process of argument parsing to remedy undesired
  dependences.  The `ArgParser` is now a member of singleton `Catalog`.
  This enables plugins to register command line arguments *before* main
  is invoked.

In addition, this patch improves the GitLab CI/CD configuration:

- Cache both static and shared libraries.
- Caching distinguishes between debug/release, shared/static builds.
- Caching includes more CMake stamp files to suppress repeated downloads
  of external projects.
- Jobs for the release builds are now run on merge requests that are
  *not* marked as WIP/Draft.