Implement `AdjacencyMatrix` as part of `util`.
The `AdjacencyMatrix` was formerly a part of `QueryGraph.hpp/cpp`.  It
also provided a c'tor to be built given a `QueryGraph` instance.

We instead now provide `AdjacencyMatrix` as part of our `util`s.  This
resolves an unnecessary coupling with `QueryGraph`.  Consequently, the
mentioned c'tor is removed.  Instead, `QueryGraph` now *owns* an
instance of `AdjacencyMatrix`.  This instance is built *on demand* when
accessed for the first time.

All lines building an `AdjacencyMatrix` from a `QueryGraph` are adapted
accordingly.