Added `Spn`.
mu*t*able now has support for the learned Datastructure Sum-Product Network
which represents a joint probability distribution.
With this, there is now the ability to estimate cardinalitites with the help
of the `SpnEstimator`.

There is an implementation in `src/util` to give the basic functionality of an
SPN and there is a wrapper of `Spn` in `src/catalog` to add the database specifics.

To implement `Spn`, further needed algorithms where implemented:

- `Kmeans`, a simple clustering algorithm
- `RDC`, a coefficient to give the dependency between non-linear random variables