Properly dispose of `v8::ArrayBuffer::Allocator`.
In the `V8Platform`, we create a `v8::ArrayBuffer::Allocator` for the
`v8::Isolate`.  The object is owned by the caller (i.e. us).  We
therefore must dispose of the object.  To do so, save the pointer to the
object in the `V8Platform` and dispose of it in the d'tor.