[API] Change API of `DatabaseInstruction`.
We did store entire `DatabaseInstruction` instances in the `Catalog`.
These instances basically just contained the instruction's runtime type
and were used for dynamic dispatching.  To pass the arguments to a
`DatabaseInstruction`, we had to provide and implement
`execute_instruction()`, taking the instruction arguments as function
parameters.

This commit changes the API of `DatabaseInstruction` as follows:

- The `Catalog` now stores `DatabaseInstructionFactory`s instead of
  `DatabaseInstruction` instances.
- A `DatabaseInstruction` now implements only `execute()` as inherited
  from `DatabaseCommand`.  Former `execute_instruction()` is removed.
- The arguments to an instruction are now passed to the
  `DatabaseInstruction` c'tor.  The arguments are stored within the
  `DatabaseInstruction` instance.
- The `Catalog` provides a factory method `create_instruction()` to
  create a fresh `DatabaseInstruction` from a provided instruction name
  and its arguments, passed as `std::vector<std::string>`.